# docs/task-center-goal-v1-openapi.yaml
openapi: 3.0.3
info:
  title: Task Center API
  version: v1
  description: |
    Task Center full handoff version:
    - P0 Goal
    - P1 Info + Tag
    - P2 DLL (minimal template + instantiate)
servers:
  - url: http://localhost:8080
tags:
  - name: Student Task Center
  - name: Teacher Task Center
paths:
  /api/student/tasks:
    get:
      tags: [Student Task Center]
      summary: List student tasks by type
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [GOAL, INFO]
        - name: status
          in: query
          required: false
          description: Goal only
          schema:
            type: string
            enum: [ALL, NOT_STARTED, IN_PROGRESS, COMPLETED]
        - name: category
          in: query
          required: false
          description: Info only
          schema:
            type: string
            enum: [ALL, ACTIVITY, VOLUNTEER]
        - name: tag
          in: query
          required: false
          description: Info only
          schema:
            type: string
        - name: keyword
          in: query
          required: false
          schema:
            type: string
        - name: unreadOnly
          in: query
          required: false
          description: Info only
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
      responses:
        '200':
          description: List by type
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/GoalListResponseDto'
                  - $ref: '#/components/schemas/InfoListResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/student/tasks/{taskId}/status:
    patch:
      tags: [Student Task Center]
      summary: Update student goal status
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalStatusRequestDto'
      responses:
        '200':
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalTaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/student/tasks/{infoId}/read:
    patch:
      tags: [Student Task Center]
      summary: Mark info as read
      parameters:
        - name: infoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
            example: {}
      responses:
        '200':
          description: Updated info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoTaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/teacher/tasks:
    get:
      tags: [Teacher Task Center]
      summary: List teacher/admin tasks by type
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [GOAL, INFO, DLL]
        - name: studentId
          in: query
          required: false
          description: Goal only
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: status
          in: query
          required: false
          description: Goal only
          schema:
            type: string
            enum: [ALL, NOT_STARTED, IN_PROGRESS, COMPLETED]
        - name: category
          in: query
          required: false
          description: Info only
          schema:
            type: string
            enum: [ALL, ACTIVITY, VOLUNTEER]
        - name: tag
          in: query
          required: false
          description: Info only
          schema:
            type: string
        - name: keyword
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
      responses:
        '200':
          description: List by type
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/GoalListResponseDto'
                  - $ref: '#/components/schemas/InfoListResponseDto'
                  - $ref: '#/components/schemas/DllTaskListResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/teacher/tasks/goals:
    post:
      tags: [Teacher Task Center]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequestDto'
            example:
              studentId: 20001
              title: 完成 OUAC 账户注册
              description: 本周内完成并截图上传
              dueAt: '2026-03-15'
      responses:
        '200':
          description: Created goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalTaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/teacher/tasks/{taskId}/status:
    patch:
      tags: [Teacher Task Center]
      summary: Update goal status (teacher/admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalStatusRequestDto'
      responses:
        '200':
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalTaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/teacher/tasks/assignable-students:
    get:
      tags: [Teacher Task Center]
      summary: List assignable students
      responses:
        '200':
          description: Student options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssignableStudentDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/teacher/tasks/infos:
    post:
      tags: [Teacher Task Center]
      summary: Publish an info task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInfoRequestDto'
            example:
              title: 义工活动报名通知
              content: 请在周五前完成报名。
              category: VOLUNTEER
              tags: [Volunteer, Grade12]
      responses:
        '200':
          description: Created info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoTaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/teacher/tasks/dll-templates:
    post:
      tags: [Teacher Task Center]
      summary: Create DLL template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDllTemplateRequestDto'
            example:
              name: OUAC 材料包模板
              description: 申请材料打包模板
              payloadSchema: '{"fields":["deadline","documents"]}'
      responses:
        '200':
          description: Created template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DllTemplateDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/teacher/tasks/dll-templates/{templateId}/instantiate:
    post:
      tags: [Teacher Task Center]
      summary: Instantiate DLL task manually
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstantiateDllTemplateRequestDto'
            example:
              assignedStudentId: 20001
              title: OUAC 材料任务
              status: IN_PROGRESS
      responses:
        '200':
          description: Created DLL task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DllTaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    GoalTaskStatus:
      type: string
      enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
    GoalTaskDto:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { type: string, enum: [GOAL] }
        title: { type: string }
        description: { type: string }
        status:
          $ref: '#/components/schemas/GoalTaskStatus'
        dueAt:
          type: string
          nullable: true
          example: '2026-03-15'
        assignedStudentId: { type: integer, format: int64 }
        assignedStudentName: { type: string }
        assignedByTeacherId: { type: integer, format: int64 }
        assignedByTeacherName: { type: string }
        createdAt: { type: string, example: '2026-03-09T11:00:00' }
        updatedAt: { type: string, example: '2026-03-09T11:00:00' }
        completedAt:
          type: string
          nullable: true
          example: '2026-03-10T09:00:00'
        progressNote: { type: string }
      required:
        - id
        - type
        - title
        - description
        - status
        - assignedStudentId
        - assignedStudentName
        - assignedByTeacherId
        - assignedByTeacherName
        - createdAt
        - updatedAt
        - progressNote
    GoalListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GoalTaskDto'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
      required: [items, total, page, size]
    AssignableStudentDto:
      type: object
      properties:
        studentId: { type: integer, format: int64 }
        studentName: { type: string }
        username: { type: string, nullable: true }
      required: [studentId, studentName]
    CreateGoalRequestDto:
      type: object
      properties:
        studentId: { type: integer, format: int64 }
        title: { type: string }
        description: { type: string }
        dueAt:
          type: string
          nullable: true
      required: [studentId, title, description]
    UpdateGoalStatusRequestDto:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/GoalTaskStatus'
        progressNote: { type: string }
      required: [status]
    InfoTaskCategory:
      type: string
      enum: [ACTIVITY, VOLUNTEER]
    InfoTaskDto:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { type: string, enum: [INFO] }
        title: { type: string }
        content: { type: string }
        category:
          $ref: '#/components/schemas/InfoTaskCategory'
        tags:
          type: array
          items:
            type: string
        targetStudentCount: { type: integer }
        publishedByTeacherId: { type: integer, format: int64 }
        publishedByTeacherName: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        read: { type: boolean }
        readAt:
          type: string
          nullable: true
      required:
        - id
        - type
        - title
        - content
        - category
        - tags
        - targetStudentCount
        - publishedByTeacherId
        - publishedByTeacherName
        - createdAt
        - updatedAt
        - read
    InfoListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InfoTaskDto'
        total: { type: integer, format: int64 }
        page: { type: integer }
        size: { type: integer }
      required: [items, total, page, size]
    CreateInfoRequestDto:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        category:
          $ref: '#/components/schemas/InfoTaskCategory'
        tags:
          type: array
          items:
            type: string
      required: [title, content, category]
    DllTaskStatus:
      type: string
      enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
    DllTemplateDto:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        payloadSchema: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, description, payloadSchema, createdAt, updatedAt]
    DllTaskDto:
      type: object
      properties:
        id: { type: integer, format: int64 }
        templateId: { type: integer, format: int64 }
        title: { type: string }
        status:
          $ref: '#/components/schemas/DllTaskStatus'
        assignedStudentId: { type: integer, format: int64 }
        createdAt: { type: string }
      required: [id, templateId, title, status, assignedStudentId, createdAt]
    DllTaskListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DllTaskDto'
        total: { type: integer, format: int64 }
        page: { type: integer }
        size: { type: integer }
      required: [items, total, page, size]
    CreateDllTemplateRequestDto:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        payloadSchema: { type: string }
      required: [name]
    InstantiateDllTemplateRequestDto:
      type: object
      properties:
        assignedStudentId: { type: integer, format: int64 }
        title: { type: string }
        status:
          $ref: '#/components/schemas/DllTaskStatus'
      required: [assignedStudentId]
    ErrorDto:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        status: { type: integer }
        details:
          type: array
          items:
            type: string
      required: [code, message, status, details]
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
          example:
            code: BAD_REQUEST
            message: status invalid
            status: 400
            details: []
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
          example:
            code: UNAUTHENTICATED
            message: Unauthenticated.
            status: 401
            details: []
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
          example:
            code: FORBIDDEN
            message: Forbidden
            status: 403
            details: []
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
          example:
            code: NOT_FOUND
            message: Resource not found.
            status: 404
            details: []