# src/main/java/com/studentmanagement/studentmanagementserver/domain/task/CreateDllTemplateRequestDto.java
package com.studentmanagement.studentmanagementserver.domain.task;

public class CreateDllTemplateRequestDto {
    private String name;
    private String description;
    private String payloadSchema;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getPayloadSchema() {
        return payloadSchema;
    }

    public void setPayloadSchema(String payloadSchema) {
        this.payloadSchema = payloadSchema;
    }
}